Notify script OnUnitAttacked only once for attacked unit, do not add self notify on unit attacked

+++ src/ai/KM_AI.pas
@@ -45,7 +45,7 @@ type
     property IsNotWinnerNotLoser: Boolean read GetIsNotWinnerNotLoser;
     procedure OwnerUpdate(aPlayer: TKMHandIndex);
     procedure HouseAttackNotification(aHouse: TKMHouse; aAttacker: TKMUnitWarrior);
-    procedure UnitAttackNotification(aUnit: TKMUnit; aAttacker: TKMUnit);
+    procedure UnitAttackNotification(aUnit: TKMUnit; aAttacker: TKMUnit; aNotifyScript: Boolean = True);
 
     procedure Save(SaveStream: TKMemoryStream);
     procedure Load(LoadStream: TKMemoryStream);
@@ -285,7 +285,7 @@ end;
 
 
 // aUnit is our unit that was attacked
-procedure TKMHandAI.UnitAttackNotification(aUnit: TKMUnit; aAttacker: TKMUnit);
+procedure TKMHandAI.UnitAttackNotification(aUnit: TKMUnit; aAttacker: TKMUnit; aNotifyScript: Boolean = True);
 const
   NotifyKind: array [Boolean] of TAttackNotification = (an_Citizens, an_Troops);
 var
@@ -328,7 +328,8 @@ begin
       end;
   end;
 
-  gScriptEvents.ProcUnitAttacked(aUnit, aAttacker); //At the end since it could kill the unit
+  if aNotifyScript then
+    gScriptEvents.ProcUnitAttacked(aUnit, aAttacker); //At the end since it could kill the unit
 end;
